% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendInterventoRMessage.R
\name{sendInterventorEMail}
\alias{sendInterventorEMail}
\title{Send a message to a proband's medical attendant via e-mail}
\usage{
sendInterventorEMail(message_content, studyXSId, probandXSID, studyApiKey,
  interventorUrl)
}
\arguments{
\item{message_content}{the content of the e-mail}

\item{studyXSId}{the xs-study id}

\item{probandXSID}{xs-proband id}

\item{studyApiKey}{the xs study api secret key}

\item{interventorUrl}{the interventor server's url}
}
\description{
Send a message to a proband's medical attendant via e-mail
}

